zhouwenwang-divination/gradio_client/requirements.txt
gradio>=4.0.0
requests>=2.31.0
sseclient-py>=1.8.0
pillow>=10.0.0
python-dotenv>=1.0.0 